+++ engine/src/gfxlib_struct.cpp
@@ -1,8 +1,12 @@
 /*
  * gfxlib_struct.cpp
- * 
- * Copyright (C) 2001-2024 Daniel Horn, pyramid3d, Stephen G. Tuggy,
- * and other Vega Strike contributors.
+ *
+ * Vega Strike - Space Simulation, Combat and Trading
+ * Copyright (C) 2001-2025 The Vega Strike Contributors:
+ * Project creator: Daniel Horn
+ * Original development team: As listed in the AUTHORS file
+ * Current development team: Roy Falk, Benjamen R. Meyer, Stephen G. Tuggy
+ *
  *
  * https://github.com/vegastrike/Vega-Strike-Engine-Source
  *
@@ -19,7 +23,7 @@
  * GNU General Public License for more details.
  *
  * You should have received a copy of the GNU General Public License
- * along with Vega Strike. If not, see <https://www.gnu.org/licenses/>.
+ * along with Vega Strike.  If not, see <https://www.gnu.org/licenses/>.
  */
 
 
@@ -134,11 +138,11 @@ void GFXVertexList::RefreshDisplayList() {
                 (changed & CHANGE_MUTABLE) ? GL_DYNAMIC_DRAW_ARB : GL_STATIC_DRAW_ARB);
         if (changed & HAS_INDEX) {
             GFXBindElementBuffer(display_list);
-            unsigned int tot = 0;
+            intmax_t tot = 0;
             for (int i = 0; i < numlists; ++i) {
                 tot += offsets[i];
             }
-            unsigned int indexsize = (changed & INDEX_BYTE)
+            intmax_t indexsize = (changed & INDEX_BYTE)
                     ? sizeof(char)
                     : ((changed & INDEX_SHORT)
                             ? sizeof(unsigned short)
